+++ backend/src/models/user.py
@@ -1,0 +1,17 @@
+from models.workout import WorkoutPlan
+
+
+from sqlmodel import Field, Relationship, SQLModel
+
+
+from typing import List, Optional
+
+
+class User(SQLModel, table=True):
+    # optional because it's generated by db. When we create the obj we don't have it
+    id: Optional[int] = Field(default=None, primary_key=True)
+    username: str = Field(index=True, unique=True)
+    email: str
+    password: str
+
+    plans: List["WorkoutPlan"] = Relationship(back_populates="creator")
